backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
PyMySQL==1.1.1
cryptography==44.0.2
pydantic==2.10.6
pydantic-settings==2.8.1
PyJWT==2.10.1
